Added Port and Ubuntu detection
I have added Ubuntu detection for the rsync install to ensure it is not missing.

I have also changed some flags in the rsync command. For the moment I have hard coded the port number as a variable which makes it easy to change , but it doesnt give the user much power over this aspect of the program

I have also updated the readme to better describe the state of next-release